+++ CHANGELOG.md
@@ -1,14 +1,13 @@
 # Change Log
 
-All notable changes to the "auto-collapse-unopened-explorer" extension will be documented
+All notable changes to the "collapse-unopened-explorer" extension will be documented
 in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how
 to structure this file.
 
-## [Unreleased]
 
 ## [1.0.0] - 2021-09-18
 
 - Initial release, forked from LukeGeneva/vscode-auto-collapse-explorer
-- Basic unoptimised version (blindly will collapse all and re-focus all)
+- Basic unoptimised version (blindly will collapse all and re-focus all), to keep it efficient it has to be activated through the command: